Improve demo project
